style : 파일이름변경 및 주석최신화, 함수이름 업데이트 - mock 서비스에 사용되는 room.ts의 파일을 room.data.ts로 바꿈 - 유틸폴더를 없애고 사용자 위치 계산기능을 기존에 사용된 util.ts 파일에 옮김 - 위치계산기능 코드에 있는 주석최신화 - 위치계산기능 코드에 존재하는 const a,c를 가독성 좋게 업데이트
